mv orlsit2inlist rule to simplifier
